refactor: update Kreuzberg benchmark configuration
- Remove redundant tesseract_default extractor
- Update documentation to reflect benchmarking results
- Benchmarking showed default configuration is already optimal
- Disabling ngram/dict correction actually makes extraction slower
- Keep default Tesseract settings for best performance

+++ src/quality_assessment.py
@@ -76,8 +76,8 @@ class TextQualityAssessor:
             "word_count": len(words),
             "sentence_count": len([s for s in sentences if s.strip()]),
             "paragraph_count": len([p for p in paragraphs if p.strip()]),
-            "avg_word_length": np.mean([len(word) for word in words]) if words else 0,
-            "avg_sentence_length": np.mean([len(s.split()) for s in sentences if s.strip()]) if sentences else 0,
+            "avg_word_length": float(np.mean([len(word) for word in words])) if words else 0.0,
+            "avg_sentence_length": float(np.mean([len(s.split()) for s in sentences if s.strip()])) if sentences else 0.0,
         }
 
     def _content_quality_metrics(self, text: str) -> dict[str, Any]:
